Fix get_object_from_id function for same name and validate data
If two kids had the same name I wasn't able to see last name. Now, if the object
has the last name property, more information is shown on the screen.
Now the user id in case that there are more than 2 coincidences inside the function
is validated. I had an error there