Fix bugs about allergic kids and show allergic menus
When a kid had more than one allergic to a recipe, they was count more than once
so the kid shoewed up twice or more in the allergic kids list. I had a mistake
in the find_kids_allergic_to_recipe.
I had another bug in the Daily_menu class, in the method to get the properties
I was taken the wrong variable in the allergic kid list so the quantity of rations
was always 1 when it had to be more than that.
I was passing the wrong parameter in the method to create a menu when there were
more than 1 allergic kid. I was passing the allergic recipe and not the main recipe